Cleaned up code, fixed some glitches, added friction to the static cars, made car properties dynamic
