+++ src/src/main/java/neonracer/model/entity/EntityCar.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import neonracer.core.GameContext;
 import neonracer.model.car.Car;
-import neonracer.phys.entity.CarPhysics;
+import neonracer.phys.entity.car.CarPhysicsFactory;
 import neonracer.render.gl.core.Texture;
 
 public class EntityCar extends Entity {
@@ -38,6 +38,8 @@ public class EntityCar extends Entity {
     @Override
     public void onInitialize(GameContext gameContext) {
         super.onInitialize(gameContext);
-        setPhysics(CarPhysics.create(gameContext, this));
+        if (getPhysics() == null)
+            setPhysics(CarPhysicsFactory.createStatic(gameContext, this));
     }
+
 }
