+++ src/src/main/java/neonracer/phys/entity/EntityPhysics.java
@@ -2,10 +2,12 @@ package neonracer.phys.entity;
 
 import org.joml.Vector2f;
 
-public abstract class EntityPhysics {
+public interface EntityPhysics {
 
-    public abstract Vector2f getPosition();
+    Vector2f getPosition();
 
-    public abstract float getRotation();
+    float getRotation();
+
+    void update();
 
 }
