+++ src/src/main/java/neonracer/util/MathHelper.java
@@ -1,0 +1,12 @@
+package neonracer.util;
+
+public class MathHelper {
+
+    public static float clamp(float f, float min, float max) {
+        if (f < min) return min;
+        if (f > max) return max;
+        return f;
+    }
+
+
+}
